+++ client/vite.config.js
@@ -1,16 +1,20 @@
-import { defineConfig } from 'vite'
+import { defineConfig, loadEnv } from 'vite'
 import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  server: {
-    host: true, // Need this for Docker
-    port: 5173,
-    proxy: {
-      '/api': {
-        target: process.env.VITE_API_URL || 'http://127.0.0.1:8001',
-        changeOrigin: true
+export default defineConfig(({ mode }) => {
+  const env = loadEnv(mode, process.cwd(), '')
+  
+  return {
+    plugins: [react()],
+    server: {
+      host: true, // Need this for Docker
+      port: 5173,
+      proxy: {
+        '/api': {
+          target: env.VITE_API_URL,
+          changeOrigin: true
+        }
       }
     }
   }